Actually stop processes gracefully
Albeit in a risky way, using a hack with process groups. These aren't
well-behaved service daemons we're supervising, they're dev-friendly
commands that expect the process they are typically waiting on to itself
exit before the outer command receives a signal, e.g. because the user
sent SIGINT to the foregrounded inner-most process in an interactive
shell.